Server/requirements.txt
pyshark
websocket-client
pyyaml